Fix warning

+++ src/oxtra/codegen/instruction.cpp
@@ -9,7 +9,7 @@ using namespace codegen::helper;
 codegen::Instruction::Instruction(const fadec::Instruction& inst, uint8_t update, uint8_t require, bool eob)
 		: fadec::Instruction{inst} {
 	update_flags = update;
-	require_flags = eob ? flags::all : require;
+	require_flags = eob ? static_cast<uint8_t>(flags::all) : require;
 	end_of_block = eob;
 }
